Support for building a Dart Docker image with a custom version of Dart
Add a script and Dockerfile for building a Docker image with a Dart SDK
from a custom .deb file.

This is convenient for testing with versions of Dart which is not available
for installation from our Debian repository.

Updated the scripts to be independent of where they are run from.

Updated the README.

R=kustermann@google.com, wibling@google.com
BUG=

Review URL: https://codereview.chromium.org//761953003